Reduced beater spawn collider specificity
Switched to a rectangle collider for the beater spawn as having a collider that exactly follows the shape is a bit pointless. The rect collider also means that you can no longer place a mover inside of the beater spawn.

However, doing this on its own creates issues with the beater as it gets pushed outside of the colliders area as soon as it spawns. To fix this, the beater spawn collider is disabled during the run. Beater spawns are also made slightly transparent when disabled to visualize this change.