Click detection & entity configuration
ClickDetector.gd is a singleton script that moves an Area2D to the mouse position every frame. This is used to get a list of entities that are at the mouse position which is in turn used for entity configuration.

The entity configuration system stores a dictionary of configuration options and their corresponding button's resource paths. Entity.gd now has a string array property which stores the available configuration options for that entity. When an entity is right-clicked on, the EntityConfigurator creates a button for each config option available for that entity. Currently, the only available config option is delete. The config menu is placed at the mouse position.