Fixed mover spawn position error
EntityManager was throwing errors as it was attempting to move the mover to the mouse position before it was added to the scene tree. This has been moved to the MoverContainer's _ready() function as this moves it as soon as all its children have been added to the scene tree.