Added bouncers
Bouncers bounce the ball in the direction that that their DirectionDisplay is pointing. Their direction can be flipped in their entity configuration menu.

The new entity configuration option 'FlipDirection' can be used to flip the direction of the bouncers. The system is set up so that it should work with all future movers that require the option.